requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
supabase==2.3.7
python-dotenv==1.0.1
PyJWT==2.10.1
gunicorn==21.2.0
